+++ src/OrleansRuntime/Streams/QueueBalancer/StreamQueueBalancerType.cs
@@ -13,23 +13,6 @@ namespace Orleans.Streams
         public static Type ConsistentRingBalancer = typeof(ConsistentRingQueueBalancer);
 
         /// <summary>
-        /// Stream queue balancer that uses Azure deployment information and silo statuses from Membership oracle for load balancing.  
-        /// Requires silo running in Azure.
-        /// This Balancer uses both the information about the full set of silos as reported by Azure role code and 
-        /// the information from Membership oracle about currently active (alive) silos and rebalances queues from non active silos.
-        /// </summary>
-        public static Type DynamicAzureDeploymentBalancer = typeof(DynamicAzureDeploymentBalancer);
-
-        /// <summary>
-        /// Stream queue balancer that uses Azure deployment information for load balancing. 
-        /// Requires silo running in Azure.
-        /// This Balancer uses both the information about the full set of silos as reported by Azure role code but 
-        /// does NOT use the information from Membership oracle about currently alive silos. 
-        /// That is, it does not rebalance queues based on dymanic changes in the cluster Membership.
-        /// </summary>
-        public static Type StaticAzureDeploymentBalancer = typeof(StaticAzureDeploymentBalancer);
-
-        /// <summary>
         /// Stream queue balancer that uses the cluster configuration to determine deployment information for load balancing.  
         /// It does not support dynamic changes to global (full) cluster configuration.
         /// This Balancer does use the information from Membership oracle about currently active (alive) silos 
@@ -50,12 +33,5 @@ namespace Orleans.Streams
         /// This balancer supports queue balancing in cluster auto-scale scenario, unexpected server failure scenario, and try to support ideal distribution 
         /// </summary>
         public static Type ClusterConfigDeploymentLeaseBasedBalancer = typeof(ClusterConfigDeploymentLeaseBasedBalancer);
-
-        /// <summary>
-        ///  Stream queue balancer that uses Azure deployment information for load balancing. 
-        /// Requires silo running in Azure.
-        /// This balancer supports queue balancing in cluster auto-scale scenario, unexpected server failure scenario, and try to support ideal distribution 
-        /// </summary>
-        public static Type AzureDeploymentLeaseBasedBalancer = typeof(AzureDeploymentLeaseBasedBalancer);
     }
 }
